pixel_crafter_gui/requirements.txt
altgraph==0.17.5
customtkinter==5.2.2
darkdetect==0.8.0
filelock==3.20.3
fsspec==2026.1.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
packaging==26.0
pefile==2024.8.26
pillow==12.1.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pywin32-ctypes==0.2.3
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
typing_extensions==4.15.0
